backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.1
uvicorn==0.34.2
